Inclusión de mensajes de guardado y borrado

+++ sistema/app/Http/Controllers/EmpleadoController.php
@@ -48,7 +48,8 @@ class EmpleadoController extends Controller
         }
 
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con éxito');
 
 
     }
@@ -110,7 +111,11 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-         Empleado::destroy($id);
-         return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+        
+        }
+         return redirect('empleado')->with('mensaje','Empleado Borrado');
     }
 }
